// src/frontend/src/backend.d.ts
import type { Principal } from "@icp-sdk/core/principal";
export interface Some<T> {
    __kind__: "Some";
    value: T;
}
export interface None {
    __kind__: "None";
}
export type Option<T> = Some<T> | None;
export interface ScoreEntry {
    name: string;
    score: bigint;
    timestamp: Time;
}
export type Time = bigint;
export interface backendInterface {
    getTopScores(): Promise<Array<ScoreEntry>>;
    saveScore(name: string, score: bigint): Promise<void>;
}